Use u64 instead of usize

+++ src/metadata.rs
@@ -33,7 +33,7 @@ pub mod trace_dump;
 #[derive(Debug)]
 pub struct MetadataStorage {
     entries: HashMap<TypeId, Box<dyn Any>>,
-    types_freqs: BTreeMap<String, usize>, // KEY = (declared_id, concrete_type)
+    types_freqs: BTreeMap<String, u64>, // KEY = (declared_id, concrete_type)
 }
 
 impl MetadataStorage {
@@ -118,7 +118,7 @@ impl MetadataStorage {
         *self.types_freqs.entry(type_id_str).or_insert(0) += 1;
     }
 
-    pub fn types_frequencies(&self) -> BTreeMap<String, usize> {
+    pub fn types_frequencies(&self) -> BTreeMap<String, u64> {
         self.types_freqs.clone()
     }
 }
